Move DeployFeeds into test helpers
