+++ integration-tests/deployment/ccip/test_helpers.go
@@ -2,11 +2,15 @@ package ccipdeployment
 
 import (
 	"context"
+	"fmt"
+	"math/big"
 	"sort"
 	"testing"
 	"time"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/aggregator_v3_interface"
+	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/generated/mock_v3_aggregator_contract"
 
 	"github.com/ethereum/go-ethereum/common"
 	chainsel "github.com/smartcontractkit/chain-selectors"
@@ -243,4 +247,56 @@ func AddLanesForAll(e deployment.Environment, state CCIPOnChainState) error {
 		}
 	}
 	return nil
+}
+
+const (
+	LINK     = "LINK"
+	WETH     = "WETH"
+	DECIMALS = 18
+)
+
+var (
+	LINK_PRICE = big.NewInt(5e18)
+)
+
+func DeployFeeds(lggr logger.Logger, chain deployment.Chain) (deployment.AddressBook, map[string]common.Address, error) {
+	ab := deployment.NewMemoryAddressBook()
+	//TODO: Maybe append LINK to the contract name
+	linkTV := deployment.NewTypeAndVersion(PriceFeed, deployment.Version1_0_0)
+	mockLinkFeed, err := deployContract(lggr, chain, ab,
+		func(chain deployment.Chain) ContractDeploy[*aggregator_v3_interface.AggregatorV3Interface] {
+			linkFeed, tx, _, err1 := mock_v3_aggregator_contract.DeployMockV3AggregatorContract(
+				chain.DeployerKey,
+				chain.Client,
+				DECIMALS,   // decimals
+				LINK_PRICE, // initialAnswer
+			)
+			aggregatorCr, err2 := aggregator_v3_interface.NewAggregatorV3Interface(linkFeed, chain.Client)
+
+			var err error
+			if err1 != nil || err2 != nil {
+				err = fmt.Errorf("linkFeedError: %w, AggregatorInterfaceError: %w", err1, err2)
+			}
+			return ContractDeploy[*aggregator_v3_interface.AggregatorV3Interface]{
+				Address: linkFeed, Contract: aggregatorCr, Tv: linkTV, Tx: tx, Err: err,
+			}
+		})
+
+	if err != nil {
+		lggr.Errorw("Failed to deploy link feed", "err", err)
+		return ab, nil, err
+	}
+
+	lggr.Infow("deployed mockLinkFeed", "addr", mockLinkFeed.Address)
+
+	desc, err := mockLinkFeed.Contract.Description(&bind.CallOpts{})
+	if err != nil {
+		lggr.Errorw("Failed to get description", "err", err)
+		return ab, nil, err
+	}
+
+	tvToAddress := map[string]common.Address{
+		desc: mockLinkFeed.Address,
+	}
+	return ab, tvToAddress, nil
 }
